+++ CHANGES.md
@@ -1,5 +1,14 @@
 # CHANGES
 
+## Release 0.1a2 (alpha)
+
+- `lock_acquire()` returns an URL as string (and not as an `RDLMLock` object)
+- `lock_release()` takes a lock URL as string (and not as an `RDLMLocki` object)
+- new methods :
+    - `lock_get()`
+    - `lock_resource_delete()`
+    - `lock_resource_delete_all()`
+
 ## Release 0.1a1 (alpha)
 
 - first release